backend/requirements.txt
blinker==1.8.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.8
dnspython==2.6.1
Flask==3.0.3
Flask-Cors==4.0.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mongoengine==0.28.2
pycparser==2.22
PyJWT==2.8.0
pymongo==4.7.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
setuptools==70.0.0
six==1.16.0
Werkzeug==3.0.3
